Split tests into multiple files (#79)
* Split tests

* removed broken merge

* Added download tests

* Added missing tests

* Fixed codacy errors

* Fixed styling issues

* More style fixes

* Removed unused imports

* Removed deletes of models at the end of tests

* init factory_boy

* changed requirements

* first test file changed

* clean up

* cleaned download tests

* fixed frontend tests

* changed script tests

* fixed websockets tests

* added factory_boy wheel

* added faker wheel

* added faker depdencies

* fixed code style

* removed manual slave creation
replaced test_script with factory

* clean
