+++ frontend/tests/test_components.py
@@ -1,0 +1,11 @@
+#  pylint: disable=C0111
+#  pylint: disable=C0103
+
+from django.test import TestCase
+
+
+class ComponentTests(TestCase):  # pylint: disable=unused-variable
+    def test_script_entry(self):
+        from frontend.templatetags.components import script_entry
+        response = script_entry("test")
+        self.assertEqual({"script": "test"}, response)
